Moved main.ts to src/index.ts
